analyzePositionEnergyRes was added and is a script that analyzes the x-scan and outputs several plots that show the detector response. fullTraceParser was the start point to parse full trace data from pixie-4s. plotBox is just to plot electrode positions on a position plot. writeEventFiles writes events from calculated WP data and can add different amounts of noise to the events.

+++ RTDataProcessing/pixieBinParser.py
@@ -286,7 +286,7 @@ class pixieParser:
                 ('ratioSec',np.float64),
                 ('t',np.uint32),
                 ('x',np.float32),
-                ('E',np.float32),
+                ('E',np.float64),
                 ('isGood',np.bool_)])
 
         eventArray = np.zeros((len(self.eventTimes,)),dtype=dt)
